Foreign Data Wrapper support
Use of foreign data wrappers speeds up data importing a lot.
There may be some bugs and incompatibilities though.
For example, zero-timestamp ("0000-00-00 00:00:00") is invalid
in postgres, so must not be present in foreign table. If there are
such records, a possible workaround is to create a view in mysql
that converts such values to NULLs and then import from this
view instead of original table.